+++ src/Framework/PaxosLease/PLeaseMsg.cpp
@@ -77,12 +77,12 @@ bool PLeaseMsg::Read(ByteString& data)
 	unsigned	nread;
 	char		*pos;
 		
-#define CheckOverflow()		if ((pos - data.buffer) >= data.length || pos < data.buffer) return false;
+#define CheckOverflow()		if ((pos - data.buffer) >= (int) data.length || pos < data.buffer) return false;
 #define ReadUint64_t(num)		(num) = strntouint64_t(pos, data.length - (pos - data.buffer), &nread); \
 								if (nread < 1) return false; pos += nread;
 #define ReadChar(c)			(c) = *pos; pos++;
 #define ReadSeparator()		if (*pos != '#') return false; pos++;
-#define ValidateLength()	if ((pos - data.buffer) != data.length) return false;
+#define ValidateLength()	if ((pos - data.buffer) != (int)data.length) return false;
 
 	pos = data.buffer;
 	CheckOverflow();
@@ -188,7 +188,7 @@ bool PLeaseMsg::Read(ByteString& data)
 
 bool PLeaseMsg::Write(ByteString& data)
 {
-	int required;
+	unsigned required;
 	
 	if		(type == PREPARE_REQUEST)
 	{
