+++ src/Framework/ReplicatedLog/ReplicatedLogMsg.cpp
@@ -19,11 +19,11 @@ bool ReplicatedLogMsg::Read(ByteString& data)
 	unsigned	nread, length;
 	char		*pos;
 	
-#define CheckOverflow()		if ((pos - data.buffer) >= data.length || pos < data.buffer) return false;
+#define CheckOverflow()		if ((pos - data.buffer) >= (int) data.length || pos < data.buffer) return false;
 #define ReadUint64_t(num)		(num) = strntouint64_t(pos, data.length - (pos - data.buffer), &nread); \
 								if (nread < 1) return false; pos += nread;
 #define ReadSeparator()		if (*pos != '|') return false; pos++;
-#define ValidateLength()	if ((pos - data.buffer) != data.length) return false;
+#define ValidateLength()	if ((pos - data.buffer) != (int) data.length) return false;
 
 	pos = data.buffer;
 	CheckOverflow();
@@ -46,7 +46,7 @@ bool ReplicatedLogMsg::Read(ByteString& data)
 
 bool ReplicatedLogMsg::Write(ByteString& data)
 {
-	int required;
+	unsigned required;
 	
 	required = snprintf(data.buffer, data.size, "%d|%" PRIu64 "|%" PRIu64 "|%d|%.*s",
 			nodeID, restartCounter, leaseEpoch,
